% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{AlaskaPrecinctBoroughMapping}
\alias{AlaskaPrecinctBoroughMapping}
\title{Mapping of Alaska precincts to boroughs / census areas, based on precinct boundaries drawn for three different periods:}
\format{A data frame with 1,335 rows and 3 variables:
\describe{
 \item{Year}{Year boundaries were in effect (see details above)}
 \item{Precinct}{The precinct identifier}
 \item{County}{the 3-digit FIPS code (i.e., minus the state prefix) for the borough or census area}
}}
\usage{
AlaskaPrecinctBoroughMapping
}
\description{
* Year=2010: Elections prior to 2012
* Year=2012: 2012 general election
* Year=2013: Elections after 2012
}
\keyword{datasets}

