% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{CountyCharacteristics}
\alias{CountyCharacteristics}
\title{Socio-economic and demographic characteristics of counties}
\format{A data frame with 3,141 rows and 54 variables:
\describe{
\item{State}{2-character FIPS code for the state}
\item{County}{5-character FIPS code for the county}
\item{MedianHouseholdIncome}{Median Household Income (2015 ACS 5-year estimate)}
\item{TotalPopulation}{Total County Population (2015 ACS 5-year estimate)}
\item{Male}{Total County Male Population (2015 ACS 5-year estimate)}
\item{Female}{County Female Population (2015 ACS 5-year estimate)}
\item{Age0_4}{County Population Age 0-4 (2015 ACS 5-year estimate)}
\item{Age5_9}{County Population Age 5-9 (2015 ACS 5-year estimate)}
\item{Age10_14}{County Population Age 10-14 (2015 ACS 5-year estimate)}
\item{Age15_19}{County Population Age 15-19 (2015 ACS 5-year estimate)}
\item{Age20_24}{County Population Age 20-24 (2015 ACS 5-year estimate)}
\item{Age25_34}{County Population Age 25-34 (2015 ACS 5-year estimate)}
\item{Age35_44}{County Population Age 35-44 (2015 ACS 5-year estimate)}
\item{Age45_54}{County Population Age 45-54 (2015 ACS 5-year estimate)}
\item{Age55_59}{County Population Age 55-59 (2015 ACS 5-year estimate)}
\item{Age60_64}{County Population Age 60-64 (2015 ACS 5-year estimate)}
\item{Age65_74}{County Population Age 65-74 (2015 ACS 5-year estimate)}
\item{Age75_84}{County Population Age 75-84 (2015 ACS 5-year estimate)}
\item{Age85}{County Population Age 85+ (2015 ACS 5-year estimate)}
\item{MedianAge}{Median Age in County (2015 ACS 5-year estimate)}
\item{White}{County Population, Race=White (2015 ACS 5-year estimate)}
\item{Black}{County Population, Race=Black (2015 ACS 5-year estimate)}
\item{AmericanIndianAlaskaNative}{County Population, Race=American Indian / Alaska Native (2015 ACS 5-year estimate)}
\item{Asian}{County Population, Race=Asian (2015 ACS 5-year estimate)}
\item{NativeHawaiianPacificIslander}{County Population, Race=Native Hawaiian / Pacific Islander (2015 ACS 5-year estimate)}
\item{OtherRace}{County Population, Race=Other (2015 ACS 5-year estimate)}
\item{Hispanic}{County Population, Ethnicity=Hispanic (2015 ACS 5-year estimate)}
\item{SimpsonDiversityIndex}{Inverse Simpson Diversity Index}
\item{Population25Plus}{County Population Age 25+ (2015 ACS 5-year estimate)}
\item{EdK8}{County Population with Education 8th grade or less (2015 ACS 5-year estimate)}
\item{Ed9_12}{County Population with Education 9th-12th grade (2015 ACS 5-year estimate)}
\item{EdHS}{County Population, High School Graduate / equivalent (2015 ACS 5-year estimate)}
\item{EdCollNoDegree}{County Population, some college (2015 ACS 5-year estimate)}
\item{EdAssocDegree}{County Population, associate degree (2015 ACS 5-year estimate)}
\item{EdBachelorDegree}{County Population, bachelor degree (2015 ACS 5-year estimate)}
\item{EdGraduateDegree}{County Population, graduate degree (2015 ACS 5-year estimate)}
\item{MedianHousingCosts}{Median County Monthly Housing Costs (2015 ACS 5-year estimate)}
\item{MfgEmp1970}{County Manufacturing Employment in 1970}
\item{MfgEmp1980}{County Manufacturing Employment in 1980}
\item{MfgEmp1990}{County Manufacturing Employment in 1990}
\item{MfgEmp2001}{County Manufacturing Employment in 2001}
\item{MfgEmp2015}{County Manufacturing Employment in 2015}
\item{TotalEmp1970}{County Total Employment in 1970}
\item{TotalEmp1980}{County Total Employment in 1980}
\item{TotalEmp1990}{County Total Employment in 1990}
\item{TotalEmp2001}{County Total Employment in 2001}
\item{TotalEmp2015}{County Total Employment in 2015}
\item{LandAreaSqMiles}{County Land Area (in square miles)}
\item{Employment}{County Employment, Oct 2015}
\item{LaborForce}{County Labor Force, Oct 2015}
\item{Unemployment}{County Unemployment, Oct 2015}
\item{NCHS_UrbanRural2013}{CDC census-based NCHS Urban-Rural Classification Scheme for Counties (2010 Census)}
\item{NCHS_UrbanRural2006}{CDC census-based NCHS Urban-Rural Classification Scheme for Counties (2000 Census)}
\item{NCHS_UrbanRural1990}{CDC census-based NCHS Urban-Rural Classification Scheme for Counties (1990 Census)}
\item{Married}{Total married population (2015 ACS 5-year estimates)}
\item{Widowed}{Total widowed population (2015 ACS 5-year estimates)}
\item{Separated}{Total separated population (2015 ACS 5-year estimates)}
\item{Divorced}{Total divorced population (2015 ACS 5-year estimates)}
\item{NeverMarried}{Total never-married population (2015 ACS 5-year estimates)}
\item{Uninsured}{Total without health insurance (2015 ACS 5-year estimates)}
\item{ForeignBorn}{Total foreign born population (2015 ACS 5-year estimates)}
\item{NonCitizen}{Total non-citizen population (2015 ACS 5-year estimates)}
\item{Disability}{Total population with disability (2015 ACS 5-year estimates)}
\item{TotalSSI}{Total persons receiving SSI benefits in 2015 (SSI Recipients by State/County)}
\item{AgedSSI}{Total persons receiving SSI aged benefits in 2015 (SSI Recipients by State/County)}
\item{BlindDisabledSSI}{Total persons receiving SSI blind/disabled benefits in 2015 (SSI Recipients by State/County)}
\item{OASDI}{Total persons receiving OASDI benefits in 2015 (SSI Recipients by State/County)}
\item{SSIPayments}{Total SSI payments received in 2015 (SSI Recipients by State/County)}
}}
\usage{
CountyCharacteristics
}
\description{
Socio-economic and demographic characteristics of counties
}
\keyword{datasets}

