% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ElectoralVotes2010}
\alias{ElectoralVotes2010}
\title{Electoral Votes by State for 2012, 2016, and 2020 elections}
\format{A data frame with 51 rows and 4 variables:
\describe{
 \item{State}{2-character FIPS code for the state}
 \item{StateAbbr}{2-character abbreviation for the state}
 \item{StateName}{Name of the state}
 \item{ElectoralVotes}{Number of electoral votes allocated to the state}
}}
\usage{
ElectoralVotes2010
}
\description{
Electoral Votes by State for 2012, 2016, and 2020 elections
}
\keyword{datasets}

