% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{PresidentialElectionResults2016}
\alias{PresidentialElectionResults2016}
\title{Presidential Election Results Dataset (2016)}
\format{A data frame with 3,141 rows and 18 variables:
\describe{
 \item{State}{2-character FIPS code for the state}
 \item{StateAbbr}{2-character abbreviation for the state}
 \item{StateName}{Name of the state}
 \item{County}{2-character FIPS code for the county}
 \item{CountyName}{Name of the county}
 \item{clinton}{Votes for Hillary Clinton}
 \item{stein}{Votes for Jill Stein}
 \item{johnson}{Votes for Gary Johnson}
 \item{trump}{Votes for Donald Trump}
 \item{other}{Votes for other candidates}
 \item{totalvotes}{Total ballots cast for president}
 \item{dPct}{Percentage of total votes for Clinton (clinton/totalvotes)}
 \item{rPct}{Percentage of total votes for Trump (trump/totalvotes)}
 \item{otherPct}{Percentage Other votes (other/totalvotes)}
 \item{dDRPct}{Clinton votes as percentage of Clinton plus Trump votes (clinton/(clinton+trump))}
 \item{rDRPct}{Trump votes as percentage of Clinton plus Trump votes (trump/(clinton+trump))}
 \item{leanD}{Lean Clinton ratio (clinton/trump)}
 \item{leanR}{Lean Trump ratio (trump/clinton)}
}}
\usage{
PresidentialElectionResults2016
}
\description{
This dataset contains county-level results of the 2016 presidential election.
}
\details{
The data in this dataset were sourced from state elections websites, but most were
obtained through the intermediary of the Steven Pettigrew dataverse dataset
at https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/MLLQDH.

Note: Alaska published results for absentee and early voting by district, not precint.
Since districts do not roll up to Boroughs (county equivalents), we allocated these
votes to precincts according to the share of each precinct's total ballots within its district, before
rolling all Alaska precincts up into their counties.
}
\keyword{datasets}

