% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{States}
\alias{States}
\title{States and certain electoral attributes}
\format{A data frame with 51 rows and 4 variables:
\describe{
 \item{State}{2-character FIPS code for the state}
 \item{StateAbbr}{2-character abbreviation for the state}
 \item{StateName}{Name of the state}
 \item{ElectoralVotes}{Number of electoral votes allocated to the state in 2010 Census for 2012, 2016, and 2020 elections}
 \item{AllowsPartyRegistration}{=1 if state allows party affiliation at voter registration, =0 otherwise}
}}
\usage{
States
}
\description{
States and certain electoral attributes
}
\keyword{datasets}

