% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{AlaskaPrecinctBoroughMapping}
\alias{AlaskaPrecinctBoroughMapping}
\title{Mapping of Alaska precincts to boroughs / census areas}
\format{A data frame with 441 rows and 2 variables:
\describe{
 \item{Precinct}{The precinct identifier}
 \item{County}{the 3-digit FIPS code (i.e., minus the state prefix) for the borough or census area}
}}
\usage{
AlaskaPrecinctBoroughMapping
}
\description{
Mapping of Alaska precincts to boroughs / census areas
}
\keyword{datasets}

